:white_check_mark: :bug: More tests and fix of bugs due to new structure
