+++ edflow/data/agnostics/concatenated.py
@@ -1,80 +1,21 @@
 from edflow.data.dataset_mixin import DatasetMixin
+from edflow.data.agnostics.subdataset import SubDataset
 import numpy as np
 
-
-class ConcatenatedDataset(DatasetMixin):
-    """A dataset which concatenates given datasets."""
-
-    def __init__(self, *datasets, balanced=False):
-        """
-        Parameters
-        ----------
-        *datasets : DatasetMixin
-            All datasets we want to concatenate
-        balanced : bool
-            If ``True`` all datasets are padded to the length of the longest
-            dataset. Padding is done in a cycled fashion.
-        """
-        self.datasets = list(datasets)
-        self.lengths = [len(d) for d in self.datasets]
-        self.boundaries = np.cumsum(self.lengths)
-        self.balanced = balanced
-        if self.balanced:
-            max_length = np.max(self.lengths)
-            for data_idx in range(len(self.datasets)):
-                data_length = len(self.datasets[data_idx])
-                if data_length != max_length:
-                    cycle_indices = [i % data_length for i in range(max_length)]
-                    self.datasets[data_idx] = SubDataset(
-                        self.datasets[data_idx], cycle_indices
-                    )
-        self.lengths = [len(d) for d in self.datasets]
-        self.boundaries = np.cumsum(self.lengths)
-
-    def get_example(self, i):
-        """Get example and add dataset index to it."""
-        did = np.where(i < self.boundaries)[0][0]
-        if did > 0:
-            local_i = i - self.boundaries[did - 1]
-        else:
-            local_i = i
-        example = self.datasets[did][local_i]
-        example["dataset_index_"] = did
-        return example
-
-    def __len__(self):
-        return sum(self.lengths)
-
-    @property
-    def labels(self):
-        # relay if data is cached
-        if not hasattr(self, "_labels"):
-            labels = dict(self.datasets[0].labels)
-            for i in range(1, len(self.datasets)):
-                new_labels = self.datasets[i].labels
-                for k in labels:
-                    labels[k] = labels[k] + new_labels[k]
-            self._labels = labels
-        return self._labels
+from edflow.data.dataset_mixin import ConcatenatedDataset
 
 
 class ExampleConcatenatedDataset(DatasetMixin):
     """Concatenates a list of datasets along the example axis.
-    .. Warning:: Docu is wrong!
-    E.g.: \n
-    \tdset1 = [{'a': 1, 'b': 2}, {'a': 3, 'b': 4}, ...] \n
-    \tdset2 = [{'a': 6, 'b': 7}, {'a': 8, 'b': 9}, ...] \n
 
-    \tdset_conc = ExampleConcatenatedDataset(dset1, dset2) \n
-    \tprint(dset_conc[0]) \n
-    \t# {'a_0': 1, 'a_1': 6, \n
-    \t#  'b_0': 2, 'b_1': 7, \n
-    \t#  'a': ['a_0', 'a_1'], \n
-    \t#  'b': ['b_0', 'b_1']} \n
+    .. note::
+        All datasets must be of same length and must return examples with the
+        same keys and behind those keys with the same type and shape.
 
-    The new keys (e.g. `a_0, a_1`) are numbered in the order they are taken
-    from datasets. Additionally, the original, shared key is preserved and
-    returns the list of newly generated keys in correct order.
+    If dataset A returns examples of form ``{'a': x, 'b': x}`` and dataset
+    B of form ``{'a': y, 'b': y}`` the ``ExampleConcatenatedDataset(A, B)``
+    return examples of form ``{'a': [x, y], 'b': [x, y]}``.
+
     """
 
     def __init__(self, *datasets):
@@ -82,8 +23,7 @@ class ExampleConcatenatedDataset(DatasetMixin):
         Parameters
         ----------
         *datasets : DatasetMixin
-            All the datasets to concatenate. Each dataset must return a dict
-            as example!
+            All the datasets to concatenate.
         """
         assert np.all(np.equal(len(datasets[0]), [len(d) for d in datasets]))
         self.datasets = datasets
